Rearrange more tests and split two tests into separate files
