+++ src/NServiceBus.AcceptanceTests/Core/BestPractices/When_sending_events_bestpractices_disabled.cs
@@ -1,4 +1,4 @@
-namespace NServiceBus.AcceptanceTests.BestPractices
+namespace NServiceBus.AcceptanceTests.Core.BestPractices
 {
     using System.Threading.Tasks;
     using AcceptanceTesting;
