+++ src/NServiceBus.AcceptanceTests/Core/BestPractices/When_subscribing_to_command.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.BestPractices
+﻿namespace NServiceBus.AcceptanceTests.Core.BestPractices
 {
     using System;
     using System.Threading.Tasks;
