+++ src/NServiceBus.AcceptanceTests/Core/BestPractices/When_subscribing_to_command_bestpractices_disabled_on_endpoint.cs
@@ -1,14 +1,14 @@
-﻿namespace NServiceBus.AcceptanceTests.BestPractices
+﻿namespace NServiceBus.AcceptanceTests.Core.BestPractices
 {
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
     using NUnit.Framework;
 
-    public class When_unsubscribing_to_command_bestpractices_disabled_on_endpoint : NServiceBusAcceptanceTest
+    public class When_subscribing_to_command_bestpractices_disabled_on_endpoint : NServiceBusAcceptanceTest
     {
         [Test]
-        public async Task Should_allow_unsubscribing_to_commands()
+        public async Task Should_allow_subscribing_to_commands()
         {
             // This test is only relevant for message driven transports since we need to use the mappings to
             // configure the publisher. The code first API would blow up unless we turn off the checks for the entire endpoint.
@@ -17,7 +17,7 @@
             Requires.MessageDrivenPubSub();
 
             var context = await Scenario.Define<ScenarioContext>()
-                .WithEndpoint<Endpoint>(b => b.When((session, c) => session.Unsubscribe<MyCommand>()))
+                .WithEndpoint<Endpoint>(b => b.When((session, c) => session.Subscribe<MyCommand>()))
                 .Done(c => c.EndpointsStarted)
                 .Run();
