+++ src/NServiceBus.AcceptanceTests/Core/Pipeline/When_sending_from_a_send_only.cs
@@ -4,7 +4,6 @@
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
-    using Features;
     using NUnit.Framework;
 
     public class When_sending_from_a_send_only : NServiceBusAcceptanceTest
@@ -24,17 +23,6 @@
             Assert.True(context.WasCalled, "The message handler should be called");
         }
 
-        [Test]
-        public async Task Should_not_need_audit_or_fault_forwarding_config_to_start()
-        {
-            var context = await Scenario.Define<Context>(c => { c.Id = Guid.NewGuid(); })
-                .WithEndpoint<SendOnlyEndpoint>()
-                .Done(c => c.SendOnlyEndpointWasStarted)
-                .Run();
-
-            Assert.True(context.SendOnlyEndpointWasStarted, "The endpoint should have started without any errors");
-        }
-
         public class Context : ScenarioContext
         {
             public bool WasCalled { get; set; }
@@ -47,41 +35,7 @@
         {
             public SendOnlyEndpoint()
             {
-                EndpointSetup<DefaultServer>(c => { c.EnableFeature<Bootstrapper>(); }).SendOnly();
-            }
-
-            public class Bootstrapper : Feature
-            {
-                public Bootstrapper()
-                {
-                    EnableByDefault();
-                }
-
-                protected override void Setup(FeatureConfigurationContext context)
-                {
-                    context.RegisterStartupTask(b => new MyTask(b.Build<Context>()));
-                }
-
-                public class MyTask : FeatureStartupTask
-                {
-                    public MyTask(Context scenarioContext)
-                    {
-                        this.scenarioContext = scenarioContext;
-                    }
-
-                    protected override Task OnStart(IMessageSession session)
-                    {
-                        scenarioContext.SendOnlyEndpointWasStarted = true;
-                        return Task.FromResult(0);
-                    }
-
-                    protected override Task OnStop(IMessageSession session)
-                    {
-                        return Task.FromResult(0);
-                    }
-
-                    readonly Context scenarioContext;
-                }
+                EndpointSetup<DefaultServer>().SendOnly();
             }
         }
