+++ src/NServiceBus.AcceptanceTests/MessageId/When_message_has_empty_id_header.cs
@@ -6,7 +6,7 @@
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
-    using Pipeline;
+    using NServiceBus.Pipeline;
     using NUnit.Framework;
 
     public class When_message_has_empty_id_header : NServiceBusAcceptanceTest
@@ -62,7 +62,7 @@
             }
         }
 
-        
+
         public class Message : IMessage
         {
         }
