+++ src/NServiceBus.AcceptanceTests/MessageId/When_message_has_no_id_header.cs
@@ -5,7 +5,7 @@
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
-    using Pipeline;
+    using NServiceBus.Pipeline;
     using NUnit.Framework;
 
     public class When_message_has_no_id_header : NServiceBusAcceptanceTest
@@ -58,7 +58,7 @@
             }
         }
 
-        
+
         public class Message : IMessage
         {
         }
