+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -56,6 +56,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Core\Feature\When_depending_on_untyped_feature.cs" />
+    <Compile Include="Core\Feature\When_registering_a_startup_task.cs" />
     <Compile Include="Core\Pipeline\When_extending_behavior_context.cs" />
     <Compile Include="Core\Conventions\When_receiving_unobtrusive_message_without_handler.cs" />
     <Compile Include="Config\When_multiple_configuration_providers.cs" />
@@ -123,8 +124,8 @@
     <Compile Include="Serialization\When_registering_additional_deserializers.cs" />
     <Compile Include="Serialization\When_no_content_type.cs" />
     <Compile Include="Core\Pipeline\When_registering_handlers_explicitly.cs" />
-    <Compile Include="Causation\When_a_message_is_faulted.cs" />
-    <Compile Include="Causation\When_a_message_is_audited.cs" />
+    <Compile Include="Pipeline\When_a_message_is_faulted.cs" />
+    <Compile Include="Pipeline\When_a_message_is_audited.cs" />
     <Compile Include="ConfigureEndpointInMemoryPersistence.cs" />
     <Compile Include="Config\When_injecting_message_session_into_handlers.cs" />
     <Compile Include="ConventionEnforcementTests.cs" />
@@ -153,10 +154,10 @@
     <Compile Include="Core\DelayedDelivery\TimeoutManager\When_TimeoutManager_is_disabled.cs" />
     <Compile Include="Mutators\When_incoming_mutator_changes_message_type.cs" />
     <Compile Include="Mutators\When_mutating.cs" />
-    <Compile Include="BestPractices\When_publishing_command_bestpractices_disabled_on_endpoint.cs" />
-    <Compile Include="BestPractices\When_sending_events_bestpractices_disabled_on_endpoint.cs" />
-    <Compile Include="BestPractices\When_sending_events_bestpractices_disabled.cs" />
-    <Compile Include="BestPractices\When_subscribing_to_command_bestpractices_disabled_on_endpoint.cs" />
+    <Compile Include="Core\BestPractices\When_publishing_command_bestpractices_disabled_on_endpoint.cs" />
+    <Compile Include="Core\BestPractices\When_sending_events_bestpractices_disabled_on_endpoint.cs" />
+    <Compile Include="Core\BestPractices\When_sending_events_bestpractices_disabled.cs" />
+    <Compile Include="Core\BestPractices\When_subscribing_to_command_bestpractices_disabled_on_endpoint.cs" />
     <Compile Include="EndpointTemplates\ConfigureExtensions.cs" />
     <Compile Include="EndpointTemplates\DefaultPublisher.cs" />
     <Compile Include="EndpointTemplates\DefaultServer.cs" />
@@ -181,13 +182,13 @@
     <Compile Include="Recoverability\Retries\When_custom_policy_does_single_delayed_retry_before_move_to_error.cs" />
     <Compile Include="Recoverability\Retries\When_custom_policy_always_moves_to_error.cs" />
     <Compile Include="Recoverability\When_error_is_overridden_in_code.cs" />
-    <Compile Include="BestPractices\When_publishing_command_bestpractices_disabled.cs" />
-    <Compile Include="BestPractices\When_unsubscribing_to_command_bestpractices_disabled_on_endpoint.cs" />
-    <Compile Include="BestPractices\When_unsubscribing_to_command.cs" />
-    <Compile Include="BestPractices\When_subscribing_to_command.cs" />
-    <Compile Include="BestPractices\When_publishing_command.cs" />
-    <Compile Include="BestPractices\When_sending_event.cs" />
-    <Compile Include="Causation\When_a_message_is_sent.cs" />
+    <Compile Include="Core\BestPractices\When_publishing_command_bestpractices_disabled.cs" />
+    <Compile Include="Core\BestPractices\When_unsubscribing_to_command_bestpractices_disabled_on_endpoint.cs" />
+    <Compile Include="Core\BestPractices\When_unsubscribing_to_command.cs" />
+    <Compile Include="Core\BestPractices\When_subscribing_to_command.cs" />
+    <Compile Include="Core\BestPractices\When_publishing_command.cs" />
+    <Compile Include="Core\BestPractices\When_sending_event.cs" />
+    <Compile Include="Pipeline\When_a_message_is_sent.cs" />
     <Compile Include="Licensing\When_a_message_is_audited.cs" />
     <Compile Include="Recoverability\When_message_with_TimeToBeReceived_fails.cs" />
     <Compile Include="Audit\When_auditing_message_with_TimeToBeReceived.cs" />
