+++ src/NServiceBus.AcceptanceTests/Sagas/When_adding_state_to_context.cs
@@ -6,7 +6,7 @@ namespace NServiceBus.AcceptanceTests.Sagas
     using EndpointTemplates;
     using Extensibility;
     using NServiceBus;
-    using Pipeline;
+    using NServiceBus.Pipeline;
     using NServiceBus.Sagas;
     using NUnit.Framework;
     using Persistence;
